feature : FeedCommand(C,U,D) swagger test
swagger 테스트

+++ src/main/java/pinup/backend/feed/command/repository/FeedRepository.java
@@ -2,8 +2,8 @@ package pinup.backend.feed.command.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import pinup.backend.feed.entity.Feed;
+import pinup.backend.feed.command.entity.Feed;
 
 @Repository
-public interface FeedRepository extends JpaRepository<Feed,Integer> {
+public interface FeedRepository extends JpaRepository<Feed, Long> {
 }
